[crater] Minor refactor to support bulk building
