+++ fontc/src/main.rs
@@ -1,16 +1,9 @@
-use std::{
-    fs::OpenOptions,
-    io::{BufWriter, Write},
-    time::Instant,
-};
+use std::{io::Write, time::Instant};
 
 use clap::Parser;
 
-use fontbe::orchestration::{AnyWorkId, Context as BeContext};
-use fontc::{
-    create_timer, init_paths, write_font_file, Args, ChangeDetector, Config, Error, JobTimer,
-};
-use fontir::orchestration::{Context as FeContext, Flags};
+use fontbe::orchestration::AnyWorkId;
+use fontc::{create_timer, Args, Error, JobTimer};
 
 fn main() {
     // catch and print errors manually, to avoid just seeing the Debug impls
@@ -48,52 +41,7 @@ fn run() -> Result<(), Error> {
         .init();
     timer.add(time.complete());
 
-    let time = create_timer(AnyWorkId::InternalTiming("Init config"), 0)
-        .queued()
-        .run();
-    let (ir_paths, be_paths) = init_paths(&args)?;
-    let config = Config::new(args)?;
-    let prev_inputs = config.init()?;
-    timer.add(time.complete());
-
-    let mut change_detector = ChangeDetector::new(
-        config.clone(),
-        ir_paths.clone(),
-        be_paths.clone(),
-        prev_inputs,
-        &mut timer,
-    )?;
-
-    let workload = fontc::create_workload(&mut change_detector, timer)?;
-
-    let fe_root = FeContext::new_root(
-        config.args.flags(),
-        ir_paths,
-        workload.current_inputs().clone(),
-    );
-    let be_root = BeContext::new_root(config.args.flags(), be_paths, &fe_root);
-    let mut timing = workload.exec(&fe_root, &be_root)?;
-
-    if config.args.flags().contains(Flags::EMIT_TIMING) {
-        let path = config.args.build_dir.join("threads.svg");
-        let out_file = OpenOptions::new()
-            .write(true)
-            .create(true)
-            .truncate(true)
-            .open(&path)
-            .map_err(|source| Error::FileIo {
-                path: path.clone(),
-                source,
-            })?;
-        let mut buf = BufWriter::new(out_file);
-        timing
-            .write_svg(&mut buf)
-            .map_err(|source| Error::FileIo { path, source })?;
-    }
-
-    change_detector.finish_successfully()?;
-
-    write_font_file(&config.args, &be_root)
+    fontc::run(args, timer)
 }
 
 fn print_verbose_version() -> Result<(), std::io::Error> {
